cleanup usage tab
